Fitur konfirmasi delete untuk data ekspedisi dan pelanggan

+++ Ekspedisi.java
@@ -161,7 +161,14 @@ class Ekspedisi {
                     editDataEkspedisi(e);
                     break;
                 case 2:
-                    hapusDataEkspedisi(e);
+                    System.out.print("Apakah anda yakin? (y/n): ");
+                    char verif = input.next().charAt(0);
+                    if (verif == 'y') {
+                        hapusDataEkspedisi(e);
+                        System.out.println(Labels.success("Data [ID: " + id + "] Berhasil Dihapus"));
+                    } else {
+                        System.out.println(Labels.error("Data [ID: " + id + "] Batal Dihapus"));
+                    }
                     pilih = 0;
                     break;
                 case 0:
